Extract live runtime and recorder modules
The runtime facade now delegates live/dry-run feed fanout and execution wiring to live.rs, and SQLx signal/order/fill persistence to recording.rs. The public run_strategy entrypoint and feature-disabled mode stubs stay in lib.rs so callers and lean builds keep the same shape.

Constraint: DB feed relocation is still a separate Phase 2 follow-up

Constraint: Preserve live DB fatality policy and live-execution feature gate

Rejected: Move database feed loading in this commit | it changes strategy-bundles ownership and deserves its own verification boundary

Confidence: high

Scope-risk: narrow

Directive: Keep new live feed/execution wiring in live.rs and recorder persistence in recording.rs

Tested: cargo check -p new-ploy-runner

Tested: scripts/check_feature_matrix.sh --quick

Tested: cargo test -p ploy-strategy-runtime --lib

Tested: scoped rustfmt --check and git diff --check

Not-tested: real live/dry-run DB connection, remote live execution, DB feed relocation
